Added Real Estate Scraper Project
This project is a 3 part project I made for softkodes / Ahmed Technologies.
The program first scrapes real estate data from several well known Australian websites using beautifulsoup4.
Afterwards the data is added to a database using SQLite. The key information were the addresses and prices.
Finally, the data is organized and displayed by using a local web server. This could be extended for use on a world wide server in the future.